refactor(stage): move CombinedStage to vendor-neutral squid/stage/composite.py
The composite went multi-vendor (PI V-308, ASI LS50) but lived in the
vendor-named pi.py, forcing the ASI turret's find_shared_ms2000 to duck-type
around an import cycle and microscope.py's ASI branch to import through the
PI module. CombinedStage now lives in squid/stage/composite.py; pi.py keeps a
compatibility re-export; find_shared_ms2000 uses a real isinstance; canonical
import sites (microscope.py, tests) migrated.
